+++ electron/preload.js
@@ -2,5 +2,5 @@ const { contextBridge, ipcRenderer } = require('electron');
 contextBridge.exposeInMainWorld('degu', {
   onPlay: (callback) => ipcRenderer.on('play', (_, data) => callback(data)),
   onStop: (callback) => ipcRenderer.on('stop', callback),
-  ended: () => ipcRenderer.send('clip-ended'),
+  ended: (playbackId) => ipcRenderer.send('clip-ended', playbackId),
 });